+++ l10n_br_nfe/models/__init__.py
@@ -2,7 +2,5 @@
 
 from . import res_partner
 from . import res_company
-#from . import nfe_document
-#from . import nfe_document_line
 from . import nfe
 from . import res_config_settings
